Refactor: move to separate files common classes: DefaultAlphabetEncoder and IdentityAlphabetEncoder
