+++ src/alphabet_encoder/default_alphabet_encoder.h
