+++ src/find_kmers_specific/find_kmers_numeric_matrix.h
@@ -4,7 +4,7 @@
 #include <Rcpp.h>
 #include <vector>
 #include "../kmer_task_config.h"
-#include "../default_alphabet_encoder.h"
+#include "../alphabet_encoder/default_alphabet_encoder.h"
 #include "../dictionary/unordered_map_wrapper.h"
 #include "../kmer_counting_result.h"
 #include "../kmer_task_solver.h"
